refactor: 미사용 로깅 변수 삭제
